move the update logic into the Smart class
